+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `calcite-alert` - `scale` is now available as a configurable attribute
 - `calcite-dropdown` now has configurable `selection-mode` (#220)
 - `no-padding` attribute for modals allowing modal content to fill space
+- `calcite-dropdown` now has configurable `type` - click or hover (#220)
 
 ### Fixed
 - Fix for `calcite-dropdown` taking up height when closed (#213)